+++ docs/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 * :release:`4.0.0 <2017-10-06>`
+* :feature:`-` Add shaking capabilities to :ref:`protocol-incubate`
 * :feature:`-` add `step_size` to dispense and dispense_full_plate methods
 * :feature:`-` add ability to specify a well as reagent source for dispense and dispense_full_plate methods
 * :feature:`-` add ability to specify `x_cassette` for dispense and dispense_full_plate methods